+++ ChangeLog.md
@@ -9,7 +9,7 @@
 
 * Infer the primary file to require from the gemspec name.
 * No longer run `bundle console`, since it is not the same as running
-  `bundle exec irb -Ilib -rfoo/bar.rb`.
+  `bundle exec irb -Ilib -rfoo/bar`.
 
 #### scm:status